独立一些函数到utools下，修复引用消息输入id为int类型，修复发送引用消息不会发送日志，修复关闭报错，为反向ws增加错误回调，修复http api发送引用消息导致的返回错误
