+++ iirosebot/serve/http_server.py
@@ -75,7 +75,7 @@ async def send_msg(request):
 
     if reply_id is not None:
         msg_id = await API.replay_msg(Message, message, private_id=private_id)
-        return msg_id
+        return return_data({"message_id": msg_id}, 'ok', 0, 200)
 
     if message_type == "private":
         msg_id = await API.send_msg_to_private(message, user_id)
@@ -124,7 +124,7 @@ async def send_private_msg(request):
     if message is not None and user_id is not None:
         if reply_id is not None:
             msg_id = await API.replay_msg(Message, message, private_id=private_id)
-            return msg_id
+            return return_data({"message_id": msg_id}, 'ok', 0, 200)
 
         msg_id = await API.send_msg_to_private(message, user_id)
         return return_data({"message_id": msg_id}, 'ok', 0, 200)
@@ -159,7 +159,7 @@ async def send_group_msg(request):
     if message is not None:
         if reply_id is not None:
             msg_id = await API.replay_msg(Message, message, private_id=private_id)
-            return msg_id
+            return return_data({"message_id": msg_id}, 'ok', 0, 200)
 
         msg_id = await API.send_msg_to_room(message)
         return return_data({"message_id": msg_id}, 'ok', 0, 200)
