+++ iirosebot/utools/__init__.py
@@ -2,3 +2,4 @@ from iirosebot.utools.compressed_gzip import *
 from iirosebot.utools.generate_token import *
 from iirosebot.utools.md5_encrypt import *
 from iirosebot.utools.uid_hex_pares import *
+from iirosebot.utools.replay_to_json import replay_to_json
