+++ iirosebot/ws_iirose/transfer_plugin.py
@@ -14,6 +14,7 @@ from iirosebot.API.api_load_config import load_config
 from iirosebot.API.decorator import function_records, MessageType
 from iirosebot.plugin_system.plugin_init import plugin_manage_data
 from iirosebot.plugin_system.plugin_transfer import plugin_transfer
+from iirosebot.utools import replay_to_json
 
 API = APIIirose()
 gold = 0
@@ -39,30 +40,6 @@ def check_start_symbols(text):
         return True, start_symbols
     else:
         return False, None
-
-
-def replay_to_json(text):
-    try:
-        text = text.split(" (hr_) ")
-        reply_list = []
-        num = 0
-        for i in text:
-            data = i.split(" (_hr) ")
-
-            if len(data) == 1:
-                reply_list[len(reply_list) - 1]['reply'] = data[0]
-                break
-
-            user_data = data[1].split("_")
-            if num == 0:
-                reply_list.append({"message": data[0], "user_name": user_data[0], "timestamp": user_data[1]})
-            else:
-                reply_list[len(reply_list) - 1]['reply'] = data[0]
-                reply_list.append({"message": data[0], "user_name": user_data[0], "timestamp": user_data[1]})
-            num += 1
-        return reply_list
-    except:
-        return []
 
 
 async def put_message_wait(event_type, data):
